backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0b1
llama-cpp-python==0.2.11
werkzeug==2.3.7
openai
